Split chart info information into its own drawable
