+++ Kumi.Game/Screens/Play/ChartLoaderDisplay.cs
@@ -1,5 +1,4 @@
 ﻿using Kumi.Game.Charts;
-using Kumi.Game.Charts.Drawables;
 using Kumi.Game.Graphics;
 using Kumi.Game.Graphics.Sprites;
 using osu.Framework.Allocation;
@@ -10,18 +9,15 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Effects;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.Sprites;
-using osu.Framework.Localisation;
-using osuTK;
 
 namespace Kumi.Game.Screens.Play;
 
 public partial class ChartLoaderDisplay : CompositeDrawable
 {
     private Container content = null!;
-    private FillFlowContainer songFlow = null!;
-    private FillFlowContainer chartFlow = null!;
     private SpriteText loadingText = null!;
     private GlowingSpriteText readyText = null!;
+    private DrawableChartInfo drawableChartInfo = null!;
 
     public readonly BindableBool IsReady = new BindableBool();
     
@@ -59,27 +55,6 @@ public partial class ChartLoaderDisplay : CompositeDrawable
                     Y = 24,
                     Children = new Drawable[]
                     {
-                        songFlow = new FillFlowContainer
-                        {
-                            Direction = FillDirection.Vertical,
-                            Spacing = new Vector2(0, 4),
-                            Anchor = Anchor.TopLeft,
-                            Origin = Anchor.TopLeft,
-                            AutoSizeAxes = Axes.Both,
-                            Children = new Drawable[]
-                            {
-                                new SpriteText
-                                {
-                                    Text = new RomanisableString(chartInfo.Metadata.Title, chartInfo.Metadata.TitleRomanised),
-                                    Font = KumiFonts.GetFont(FontFamily.Montserrat, FontWeight.Medium, 24)
-                                },
-                                new SpriteText
-                                {
-                                    Text = new RomanisableString(chartInfo.Metadata.Artist, chartInfo.Metadata.ArtistRomanised),
-                                    Font = KumiFonts.GetFont(FontFamily.Montserrat, FontWeight.Light, 20)
-                                },
-                            }
-                        },
                         loadingText = new SpriteText
                         {
                             Anchor = Anchor.TopCentre,
@@ -98,51 +73,7 @@ public partial class ChartLoaderDisplay : CompositeDrawable
                             Alpha = 0,
                             AlwaysPresent = true,
                         },
-                        chartFlow = new FillFlowContainer
-                        {
-                            Direction = FillDirection.Vertical,
-                            Spacing = new Vector2(0, 4),
-                            Anchor = Anchor.TopRight,
-                            Origin = Anchor.TopRight,
-                            AutoSizeAxes = Axes.Both,
-                            Children = new Drawable[]
-                            {
-                                new FillFlowContainer
-                                {
-                                    Direction = FillDirection.Horizontal,
-                                    Spacing = new Vector2(4, 0),
-                                    AutoSizeAxes = Axes.Both,
-                                    Children = new Drawable[]
-                                    {
-                                        new Container
-                                        {
-                                            AutoSizeAxes = Axes.X,
-                                            RelativeSizeAxes = Axes.Y,
-                                            Children = new Drawable[]
-                                            {
-                                                new LoaderDifficultyPill(0.0f)
-                                                {
-                                                    Anchor = Anchor.CentreLeft,
-                                                    Origin = Anchor.CentreLeft,
-                                                },
-                                            }
-                                        },
-                                        new SpriteText
-                                        {
-                                            Text = chartInfo.DifficultyName,
-                                            Font = KumiFonts.GetFont(FontFamily.Montserrat, FontWeight.Medium, 20)
-                                        },      
-                                    }
-                                },
-                                new SpriteText
-                                {
-                                    Text = chartInfo.ChartSet?.Creator?.Username ?? "Unknown creator",
-                                    Font = KumiFonts.GetFont(FontFamily.Montserrat, FontWeight.Light),
-                                    Anchor = Anchor.TopRight,
-                                    Origin = Anchor.TopRight
-                                },
-                            }
-                        },
+                        drawableChartInfo = new DrawableChartInfo(chartInfo)
                     }
                 }
             }
@@ -159,53 +90,11 @@ public partial class ChartLoaderDisplay : CompositeDrawable
                 return;
 
             content.ResizeWidthTo(1.1f, 500, Easing.OutQuint);
-            songFlow.FadeOut(500, Easing.OutQuint);
-            chartFlow.FadeOut(500, Easing.OutQuint);
+            drawableChartInfo.FadeOut(500, Easing.OutQuint);
 
             loadingText.FadeOut(500, Easing.OutQuint);
             readyText.FadeIn(500, Easing.OutQuint);
         }, true);
-    }
-    
-    private partial class LoaderDifficultyPill : DifficultyPill
-    {
-        public LoaderDifficultyPill(float difficulty)
-            : base(difficulty)
-        {
-        }
-
-        protected override Drawable CreateContent()
-            => new FillFlowContainer
-            {
-                Direction = FillDirection.Horizontal,
-                Spacing = new Vector2(4, 0),
-                Colour = Colours.Gray(0.1f),
-                Anchor = Anchor.Centre,
-                Origin = Anchor.Centre,
-                AutoSizeAxes = Axes.Both,
-                Padding = new MarginPadding
-                {
-                    Vertical = 1,
-                    Horizontal = 8
-                },
-                Children = new Drawable[]
-                {
-                    new SpriteIcon
-                    {
-                        Icon = FontAwesome.Solid.Star,
-                        Size = new Vector2(12),
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre
-                    },
-                    new SpriteText
-                    {
-                        Text = Difficulty.ToString("N2"),
-                        Font = KumiFonts.GetFont(FontFamily.Montserrat, FontWeight.SemiBold, 12),
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre
-                    }
-                }
-            };
     }
 
     private partial class LoaderBar : Container
